+++ cider-inspector.el
@@ -31,12 +31,11 @@
 (require 'easymenu)
 (require 'seq)
 (require 'cider-eval)
+(require 'cider-util)
 
 ;; ===================================
 ;; Inspector Key Map and Derived Mode
 ;; ===================================
-
-(defconst cider-inspector-buffer "*cider-inspect*")
 
 ;;; Customization
 (defgroup cider-inspector nil